+++ Type/TypeInterface.php
@@ -3,8 +3,7 @@
 namespace Tdn\PhpTypes\Type;
 
 /**
- * Interface TypeInterface
- * @package Tdn\PhpTypes\Type
+ * Interface TypeInterface.
  */
 interface TypeInterface
 {
@@ -12,9 +11,10 @@ interface TypeInterface
      * A little bit of java goodness. Should be useful in the upcoming php7.
      *
      * Returns an instance of the implementing class with it's value evaluated from the argument.
-     * E.g. Boolean->valueOf("false") returns new Boolean(false).
+     * E.g. BooleanType->valueOf("false") returns new BooleanType(false).
      *
      * @param mixed $mixed
+     *
      * @return static
      */
     public static function valueOf($mixed);